requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
crewai